2/7/2018 12:16am
Was having problems with the log fold change plot. Tracked to having
duplicate glommed taxa when condensed to two ranks (high and low).

These were cropping up for two reasons:

1. NAs in lower ranks
2. Rare varieties that had the same name at a certain level
   (such as Family) despite having a different name at an
   intermeidate level such as (Order).

For the first, changed getAggregatedTaxaRankCounts to filter for NAs
at the lowest level.

For the second, created the shortened glommed_taxa, but set the levels
to the full glommed_taxa so there would be no duplicates.

Perhaps could do away with the first step as might throw out good ASVs